requirements.txt
Flask
rembg
Pillow
gunicorn # Servidor WSGI para produção
